feat(ingest): add Airflow options
