+++ metadata-ingestion/src/datahub_provider/lineage/datahub.py
@@ -1,207 +1,105 @@
 import json
 from typing import TYPE_CHECKING, Dict, List, Optional
 
-import dateutil.parser
 from airflow.configuration import conf
 from airflow.lineage.backend import LineageBackend
 
-import datahub.emitter.mce_builder as builder
-import datahub.metadata.schema_classes as models
+from datahub_provider._lineage_core import (
+    DatahubBasicLineageConfig,
+    send_lineage_to_datahub,
+)
 
 if TYPE_CHECKING:
-    from airflow import DAG
     from airflow.models.baseoperator import BaseOperator
 
-    from datahub_provider.hooks.datahub import DatahubGenericHook
+
+class DatahubLineageConfig(DatahubBasicLineageConfig):
+    # If set to true, most runtime errors in the lineage backend will be
+    # suppressed and will not cause the overall task to fail. Note that
+    # configuration issues will still throw exceptions.
+    graceful_exceptions: bool = True
 
 
-def _entities_to_urn_list(iolets: List) -> List[str]:
-    return [let.urn for let in iolets]
+def get_lineage_config() -> DatahubLineageConfig:
+    """Load the lineage config from airflow.cfg."""
 
+    # The kwargs pattern is also used for secret backends.
+    kwargs_str = conf.get("lineage", "datahub_kwargs", fallback="{}")
+    kwargs = json.loads(kwargs_str)
 
-def make_emitter_hook() -> "DatahubGenericHook":
-    # This is necessary to avoid issues with circular imports.
-    from datahub_provider.hooks.datahub import DatahubGenericHook
+    # Continue to support top-level datahub_conn_id config.
+    datahub_conn_id = conf.get("lineage", "datahub_conn_id", fallback=None)
+    if datahub_conn_id:
+        kwargs["datahub_conn_id"] = datahub_conn_id
 
-    _datahub_conn_id = conf.get("lineage", "datahub_conn_id")
-    return DatahubGenericHook(_datahub_conn_id)
+    return DatahubLineageConfig.parse_obj(kwargs)
 
 
 class DatahubLineageBackend(LineageBackend):
+    def __init__(self) -> None:
+        super().__init__()
+
+        # By attempting to get and parse the config, we can detect configuration errors
+        # ahead of time. The init method is only called in Airflow 2.x.
+        _ = get_lineage_config()
+
     # With Airflow 2.0, this can be an instance method. However, with Airflow 1.10.x, this
     # method is used statically, even though LineageBackend declares it as an instance variable.
     @staticmethod
     def send_lineage(
         operator: "BaseOperator",
-        inlets: Optional[List] = None,
-        outlets: Optional[List] = None,
+        inlets: Optional[List] = None,  # unused
+        outlets: Optional[List] = None,  # unused
         context: Dict = None,
     ) -> None:
-        # This is necessary to avoid issues with circular imports.
-        from airflow.lineage import prepare_lineage
-        from airflow.serialization.serialized_objects import (
-            SerializedBaseOperator,
-            SerializedDAG,
-        )
+        config = get_lineage_config()
 
-        from datahub_provider.hooks.datahub import AIRFLOW_1
+        try:
+            # This is necessary to avoid issues with circular imports.
+            from airflow.lineage import prepare_lineage
 
-        # Detect Airflow 1.10.x inlet/outlet configurations in Airflow 2.x, and
-        # convert to the newer version. This code path will only be triggered
-        # when 2.x receives a 1.10.x inlet/outlet config.
-        needs_repeat_preparation = False
-        if (
-            not AIRFLOW_1
-            and isinstance(operator._inlets, list)
-            and len(operator._inlets) == 1
-            and isinstance(operator._inlets[0], dict)
-        ):
-            from airflow.lineage import AUTO
+            from datahub_provider.hooks.datahub import AIRFLOW_1
 
-            operator._inlets = [
-                # See https://airflow.apache.org/docs/apache-airflow/1.10.15/lineage.html.
-                *operator._inlets[0].get(
-                    "datasets", []
-                ),  # assumes these are attr-annotated
-                *operator._inlets[0].get("task_ids", []),
-                *([AUTO] if operator._inlets[0].get("auto", False) else []),
-            ]
-            needs_repeat_preparation = True
-        if (
-            not AIRFLOW_1
-            and isinstance(operator._outlets, list)
-            and len(operator._outlets) == 1
-            and isinstance(operator._outlets[0], dict)
-        ):
-            operator._outlets = [*operator._outlets[0].get("datasets", [])]
-            needs_repeat_preparation = True
-        if needs_repeat_preparation:
-            # Rerun the lineage preparation routine, now that the old format has been translated to the new one.
-            prepare_lineage(lambda self, ctx: None)(operator, context)
+            # Detect Airflow 1.10.x inlet/outlet configurations in Airflow 2.x, and
+            # convert to the newer version. This code path will only be triggered
+            # when 2.x receives a 1.10.x inlet/outlet config.
+            needs_repeat_preparation = False
+            if (
+                not AIRFLOW_1
+                and isinstance(operator._inlets, list)
+                and len(operator._inlets) == 1
+                and isinstance(operator._inlets[0], dict)
+            ):
+                from airflow.lineage import AUTO
 
-        context = context or {}  # ensure not None to satisfy mypy
+                operator._inlets = [
+                    # See https://airflow.apache.org/docs/apache-airflow/1.10.15/lineage.html.
+                    *operator._inlets[0].get(
+                        "datasets", []
+                    ),  # assumes these are attr-annotated
+                    *operator._inlets[0].get("task_ids", []),
+                    *([AUTO] if operator._inlets[0].get("auto", False) else []),
+                ]
+                needs_repeat_preparation = True
+            if (
+                not AIRFLOW_1
+                and isinstance(operator._outlets, list)
+                and len(operator._outlets) == 1
+                and isinstance(operator._outlets[0], dict)
+            ):
+                operator._outlets = [*operator._outlets[0].get("datasets", [])]
+                needs_repeat_preparation = True
+            if needs_repeat_preparation:
+                # Rerun the lineage preparation routine, now that the old format has been translated to the new one.
+                prepare_lineage(lambda self, ctx: None)(operator, context)
 
-        dag: "DAG" = context["dag"]
-        task = context["task"]
-
-        # TODO: capture context
-        # context dag_run
-        # task_instance: "TaskInstance" = context["task_instance"]
-        # TODO: capture raw sql from db operators
-
-        flow_urn = builder.make_data_flow_urn("airflow", dag.dag_id)
-        job_urn = builder.make_data_job_urn_with_flow(flow_urn, task.task_id)
-
-        base_url = conf.get("webserver", "base_url")
-        flow_url = f"{base_url}/tree?dag_id={dag.dag_id}"
-        job_url = f"{base_url}/taskinstance/list/?flt1_dag_id_equals={dag.dag_id}&_flt_3_task_id={task.task_id}"
-        # operator.log.info(f"{flow_url=}")
-        # operator.log.info(f"{job_url=}")
-        # operator.log.info(f"{dag.get_serialized_fields()=}")
-        # operator.log.info(f"{task.get_serialized_fields()=}")
-        # operator.log.info(f"{SerializedDAG.serialize_dag(dag)=}")
-
-        flow_property_bag: Dict[str, str] = {
-            key: repr(value)
-            for (key, value) in SerializedDAG.serialize_dag(dag).items()
-        }
-        for key in dag.get_serialized_fields():
-            if key not in flow_property_bag:
-                flow_property_bag[key] = repr(getattr(dag, key))
-        job_property_bag: Dict[str, str] = {
-            key: repr(value)
-            for (key, value) in SerializedBaseOperator.serialize_operator(task).items()
-        }
-        for key in task.get_serialized_fields():
-            if key not in job_property_bag:
-                job_property_bag[key] = repr(getattr(task, key))
-        # operator.log.info(f"{flow_property_bag=}")
-        # operator.log.info(f"{job_property_bag=}")
-
-        timestamp = int(dateutil.parser.parse(context["ts"]).timestamp() * 1000)
-        ownership = models.OwnershipClass(
-            owners=[
-                models.OwnerClass(
-                    owner=builder.make_user_urn(dag.owner),
-                    type=models.OwnershipTypeClass.DEVELOPER,
-                    source=models.OwnershipSourceClass(
-                        type=models.OwnershipSourceTypeClass.SERVICE,
-                        url=dag.filepath,
-                    ),
-                )
-            ],
-            lastModified=models.AuditStampClass(
-                time=timestamp, actor=builder.make_user_urn("airflow")
-            ),
-        )
-        # operator.log.info(f"{ownership=}")
-
-        tags = models.GlobalTagsClass(
-            tags=[
-                models.TagAssociationClass(tag=builder.make_tag_urn(f"airflow_{tag}"))
-                for tag in (dag.tags or [])
-            ]
-        )
-        # operator.log.info(f"{tags=}")
-
-        flow_mce = models.MetadataChangeEventClass(
-            proposedSnapshot=models.DataFlowSnapshotClass(
-                urn=flow_urn,
-                aspects=[
-                    models.DataFlowInfoClass(
-                        name=dag.dag_id,
-                        description=f"{dag.description}\n\n{dag.doc_md or ''}",
-                        customProperties=flow_property_bag,
-                        externalUrl=flow_url,
-                    ),
-                    ownership,
-                    tags,
-                ],
+            context = context or {}  # ensure not None to satisfy mypy
+            send_lineage_to_datahub(
+                config, operator, operator.inlets, operator.outlets, context
             )
-        )
-
-        job_mce = models.MetadataChangeEventClass(
-            proposedSnapshot=models.DataJobSnapshotClass(
-                urn=job_urn,
-                aspects=[
-                    models.DataJobInfoClass(
-                        name=task.task_id,
-                        type=models.AzkabanJobTypeClass.COMMAND,
-                        description=None,
-                        customProperties=job_property_bag,
-                        externalUrl=job_url,
-                    ),
-                    models.DataJobInputOutputClass(
-                        inputDatasets=_entities_to_urn_list(inlets or []),
-                        outputDatasets=_entities_to_urn_list(outlets or []),
-                    ),
-                    ownership,
-                    tags,
-                ],
-            )
-        )
-
-        force_entity_materialization = [
-            models.MetadataChangeEventClass(
-                proposedSnapshot=models.DatasetSnapshotClass(
-                    urn=iolet,
-                    aspects=[
-                        models.StatusClass(removed=False),
-                    ],
-                )
-            )
-            for iolet in _entities_to_urn_list((inlets or []) + (outlets or []))
-        ]
-
-        hook = make_emitter_hook()
-
-        mces = [
-            flow_mce,
-            job_mce,
-            *force_entity_materialization,
-        ]
-        operator.log.info(
-            "DataHub lineage backend - emitting metadata:\n"
-            + "\n".join(json.dumps(mce.to_obj()) for mce in mces)
-        )
-        hook.emit_mces(mces)
+        except Exception as e:
+            if config.graceful_exceptions:
+                operator.log.error(e)
+                operator.log.info("Supressing error because graceful_exceptions is set")
+            else:
+                raise
