+++ metadata-ingestion/tests/unit/test_airflow.py
@@ -174,6 +174,9 @@ def test_lineage_backend(mock_emit, inlets, outlets):
         {
             "AIRFLOW__LINEAGE__BACKEND": "datahub_provider.lineage.datahub.DatahubLineageBackend",
             "AIRFLOW__LINEAGE__DATAHUB_CONN_ID": datahub_rest_connection_config.conn_id,
+            "AIRFLOW__LINEAGE__DATAHUB_KWARGS": json.dumps(
+                {"graceful_exceptions": False}
+            ),
         },
     ), mock.patch("airflow.models.BaseOperator.xcom_pull", autospec=True), mock.patch(
         "airflow.models.BaseOperator.xcom_push", autospec=True
